ai-service/requirements.txt
Flask==2.0.3
flask-cors==3.0.10
openai==0.27.0
python-dotenv==0.19.0
requests==2.28.1
gunicorn==20.1.0
Werkzeug==2.2.2